organised the folders,files and divided code into multiple components
